refactor: 검색 코드 구조 수정
