+++ src/constants/location.ts
@@ -1,0 +1,4 @@
+export const SEARCH_OPTION = {
+    ORIGIN: 'origin',
+    DESTINATION: 'destination',
+};
