+++ src/features/location/search/Search.tsx
@@ -1,9 +1,8 @@
-import { ChangeEvent, useCallback, useEffect, useState } from 'react';
+import { ChangeEvent, Dispatch, useCallback, useEffect, useState } from 'react';
 
 import LocationPointerIcon from '#/assets/icons/location-pointer.svg?react';
 import { Header } from '#/components/header';
 import useIntersectionObserver from '#/hooks/useIntersectionObserver';
-import { HeaderContents } from '#/pages/templates/header-contents';
 import { useGetLocationSearch } from '#/query-hooks/location/query';
 import {
     GeolocationCoordinatesType,
@@ -12,7 +11,6 @@ import {
 } from '#/types/location';
 import { debounce } from '#/utils/common';
 
-import { CurrentLocationSearch } from '../current-location-search';
 import { SearchBox } from '../search-box/SearchBox';
 import { SearchInput } from '../search-input/SearchInput';
 import { SearchTip } from '../search-tip/SearchTip';
@@ -22,12 +20,14 @@ import * as S from './Search.style';
 interface SearchListPropsType {
     onSelect: (location: SelectedLocationType) => void;
     onClose: () => void;
+    setCurrentLocationSearch: Dispatch<boolean>;
 }
 
-export function Search({ onSelect, onClose }: SearchListPropsType) {
-    const [isCurrentLocationSearch, setIsCurrentLocationSearch] =
-        useState<boolean>(false);
-
+export function Search({
+    onSelect,
+    onClose,
+    setCurrentLocationSearch,
+}: SearchListPropsType) {
     const [searchInput, setSearchInput] = useState<string>('');
     const [searchKeyword, setSearchKeyword] = useState<string>('');
     const [geolocationCoordinates, setGeolocationCoordinates] =
@@ -79,99 +79,64 @@ export function Search({ onSelect, onClose }: SearchListPropsType) {
 
     const handleOpenCurrentLocationSearch = () => {
         setSearchInput('');
-        setIsCurrentLocationSearch(true);
-    };
-
-    const handleCloseCurrentLocationSearch = () => {
-        setIsCurrentLocationSearch(false);
+        setCurrentLocationSearch(true);
     };
 
     return (
         <S.Wrapper>
-            {isCurrentLocationSearch ? (
-                <CurrentLocationSearch
-                    onSelect={onSelect}
-                    onCloseSearch={handleCloseCurrentLocationSearch}
-                />
-            ) : (
-                <>
-                    <Header
-                        title="위치 검색"
-                        isVisibleClose
-                        onClickClose={onClose}
+            <Header title="위치 검색" isVisibleClose onClickClose={onClose} />
+
+            <S.ContentsContainer>
+                <S.HeaderContents>
+                    <SearchInput
+                        value={searchInput}
+                        isInvalid={hasNoResult}
+                        isValid={hasResult}
+                        placeholder="지번, 도로명, 건물명으로 검색"
+                        onChangeValue={handleChangeSearchInput}
+                        onReset={() => setSearchInput('')}
                     />
-                    <HeaderContents>
-                        <S.ContentsContainer>
-                            <S.HeaderContents>
-                                <SearchInput
-                                    value={searchInput}
-                                    isInvalid={hasNoResult}
-                                    isValid={hasResult}
-                                    placeholder="지번, 도로명, 건물명으로 검색"
-                                    onChangeValue={handleChangeSearchInput}
-                                    onReset={() => setSearchInput('')}
+
+                    <S.CurrentLocation
+                        onClick={handleOpenCurrentLocationSearch}
+                    >
+                        <S.LocationPointerContainer>
+                            <LocationPointerIcon width={12} height={12} />
+                        </S.LocationPointerContainer>
+                        <S.CurrentLocationText>
+                            현재 위치로 주소 찾기
+                        </S.CurrentLocationText>
+                    </S.CurrentLocation>
+
+                    <S.DashedBorder />
+                </S.HeaderContents>
+
+                <S.ListContents>
+                    {(isEmptyInput || hasNoResult) && (
+                        <S.SearchTipContainer>
+                            <SearchTip>
+                                {isEmptyInput
+                                    ? '위치를 입력해주세요.'
+                                    : '검색 결과가 없습니다.'}
+                            </SearchTip>
+                        </S.SearchTipContainer>
+                    )}
+
+                    {isSuccess && (
+                        <S.SearchList>
+                            {data.map((item: LocationInformationType) => (
+                                <SearchBox
+                                    key={item.pkey}
+                                    searchInput={searchInput}
+                                    location={item}
+                                    onSelect={onSelect}
                                 />
-
-                                <S.CurrentLocation
-                                    onClick={handleOpenCurrentLocationSearch}
-                                >
-                                    <S.LocationPointerContainer>
-                                        <LocationPointerIcon
-                                            width={12}
-                                            height={12}
-                                        />
-                                    </S.LocationPointerContainer>
-                                    <S.CurrentLocationText>
-                                        현재 위치로 주소 찾기
-                                    </S.CurrentLocationText>
-                                </S.CurrentLocation>
-
-                                <S.SearchContainer>
-                                    {hasResult && (
-                                        <>
-                                            <S.SearchText>
-                                                <S.Keyword>
-                                                    {searchKeyword}
-                                                </S.Keyword>{' '}
-                                                검색된 주소
-                                            </S.SearchText>
-                                            <S.DashedBorder />
-                                        </>
-                                    )}
-                                </S.SearchContainer>
-                            </S.HeaderContents>
-
-                            <S.ListContents>
-                                {(isEmptyInput || hasNoResult) && (
-                                    <S.SearchTipContainer>
-                                        <SearchTip>
-                                            {isEmptyInput
-                                                ? '위치를 입력해주세요.'
-                                                : '검색 결과가 없습니다.'}
-                                        </SearchTip>
-                                    </S.SearchTipContainer>
-                                )}
-
-                                {isSuccess && (
-                                    <S.SearchList>
-                                        {data.map(
-                                            (item: LocationInformationType) => (
-                                                <SearchBox
-                                                    key={item.pkey}
-                                                    searchInput={searchInput}
-                                                    location={item}
-                                                    onSelect={onSelect}
-                                                />
-                                            ),
-                                        )}
-                                    </S.SearchList>
-                                )}
-                                <div ref={targetRef}></div>
-                            </S.ListContents>
-                        </S.ContentsContainer>
-                    </HeaderContents>
-                </>
-            )}
+                            ))}
+                        </S.SearchList>
+                    )}
+                    <div ref={targetRef}></div>
+                </S.ListContents>
+            </S.ContentsContainer>
         </S.Wrapper>
     );
 }
