--- src/pages/templates/header-contents/HeaderContents.style.ts
@@ -1,7 +1,0 @@
-import styled from '@emotion/styled';
-
-export const Container = styled.div`
-    padding-top: 64px;
-    height: 100%;
-    box-sizing: border-box;
-`;
