--- src/pages/templates/header-contents/HeaderContents.tsx
@@ -1,7 +1,0 @@
-import { PropsWithChildren } from 'react';
-
-import * as S from './HeaderContents.style';
-
-export function HeaderContents({ children }: PropsWithChildren) {
-    return <S.Container>{children}</S.Container>;
-}
