--- src/pages/templates/header-contents/index.ts
@@ -1,1 +1,0 @@
-export { HeaderContents } from './HeaderContents';
